chore: move filter & log into internal
