+++ api/internal/core/storage/etcd.go
@@ -24,8 +24,8 @@ import (
 	"go.etcd.io/etcd/clientv3"
 
 	"github.com/apisix/manager-api/conf"
+	"github.com/apisix/manager-api/internal/log"
 	"github.com/apisix/manager-api/internal/utils"
-	"github.com/apisix/manager-api/log"
 )
 
 var (
