+++ api/internal/core/store/validate.go
@@ -27,7 +27,7 @@ import (
 
 	"github.com/apisix/manager-api/conf"
 	"github.com/apisix/manager-api/internal/core/entity"
-	"github.com/apisix/manager-api/log"
+	"github.com/apisix/manager-api/internal/log"
 )
 
 type Validator interface {
