+++ api/internal/filter/authentication.go
@@ -20,11 +20,11 @@ import (
 	"net/http"
 	"strings"
 
-	"github.com/dgrijalva/jwt-go"
 	"github.com/gin-gonic/gin"
 
 	"github.com/apisix/manager-api/conf"
-	"github.com/apisix/manager-api/log"
+	"github.com/apisix/manager-api/internal/log"
+	jwt "github.com/dgrijalva/jwt-go"
 )
 
 func Authentication() gin.HandlerFunc {
