+++ api/internal/route.go
@@ -27,7 +27,7 @@ import (
 	"github.com/gin-gonic/gin"
 
 	"github.com/apisix/manager-api/conf"
-	"github.com/apisix/manager-api/filter"
+	"github.com/apisix/manager-api/internal/filter"
 	"github.com/apisix/manager-api/internal/handler"
 	"github.com/apisix/manager-api/internal/handler/authentication"
 	"github.com/apisix/manager-api/internal/handler/consumer"
@@ -40,7 +40,7 @@ import (
 	"github.com/apisix/manager-api/internal/handler/service"
 	"github.com/apisix/manager-api/internal/handler/ssl"
 	"github.com/apisix/manager-api/internal/handler/upstream"
-	"github.com/apisix/manager-api/log"
+	"github.com/apisix/manager-api/internal/log"
 )
 
 func SetUpRouter() *gin.Engine {
